requirements.txt
langchain
langchain-community
langchain-siliconflow
chromadb
python-dotenv
pypdf
python-docx
langchain-text-splitters